refactor(common): 优化代码结构和性能
- 优化 RedissonConfig 中的配置方式
-改进 JsonConverter 中的异常处理
- 为 AsyncOffsetState 类添加 EqualsAndHashCode 注解
-重构 DateFormatTransform 中的时间格式处理- 完善 TimestampTransform 中的枚举处理
- 优化 CastUtil 中的类型转换逻辑
- 改进 ConfigKeys 类中的代码结构
- 简化 Template 类中的断言逻辑
-优化 InterpolatedStringValidator 中的变量处理
- 提升 MapValidator 中的性能
- 修正 TopicNameValidator 中的错误信息
- 优化 Validators 中的字符串处理

+++ common/kafka-connector-common/src/main/java/xyz/kafka/connector/validator/TopicNameValidator.java
@@ -25,7 +25,7 @@ public class TopicNameValidator extends Validators.SingleOrListValidator {
         if (topicName.isEmpty()) {
             throw new ConfigException(name, value, "topic names may not be empty");
         } else if (topicName.length() > 249) {
-            throw new ConfigException(name, value, String.format("topic names may not be longer than %d characters", Integer.valueOf((int) MAX_LENGTH)));
+            throw new ConfigException(name, value, String.format("topic names may not be longer than %d characters", MAX_LENGTH));
         } else if (!PATTERN.matcher(value.toString()).matches()) {
             throw new ConfigException(name, value, "topic names may have 1-249 ASCII alphanumeric, `+`, `.`, `_`, and `-` characters");
         }
